refactor(cli): split watch command facade
